requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-filter==24.3
idna==3.10
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
